renamed commands/user to commands/userInformation
